fixRegression(map,geo): Resolve circular dependencies introduced by this version.
